LayoutTests/imported/w3c/web-platform-tests/css/css-text/animations/hyphen-no-interpolation-expected.txt


PASS CSS Transitions: property <hyphenate-character> from [initial] to ["e"] at (-0.3) should be ["e"]
PASS CSS Transitions: property <hyphenate-character> from [initial] to ["e"] at (0) should be ["e"]
PASS CSS Transitions: property <hyphenate-character> from [initial] to ["e"] at (0.3) should be ["e"]
PASS CSS Transitions: property <hyphenate-character> from [initial] to ["e"] at (0.5) should be ["e"]
PASS CSS Transitions: property <hyphenate-character> from [initial] to ["e"] at (0.6) should be ["e"]
PASS CSS Transitions: property <hyphenate-character> from [initial] to ["e"] at (1) should be ["e"]
PASS CSS Transitions: property <hyphenate-character> from [initial] to ["e"] at (1.5) should be ["e"]
PASS CSS Transitions with transition: all: property <hyphenate-character> from [initial] to ["e"] at (-0.3) should be ["e"]
PASS CSS Transitions with transition: all: property <hyphenate-character> from [initial] to ["e"] at (0) should be ["e"]
PASS CSS Transitions with transition: all: property <hyphenate-character> from [initial] to ["e"] at (0.3) should be ["e"]
PASS CSS Transitions with transition: all: property <hyphenate-character> from [initial] to ["e"] at (0.5) should be ["e"]
PASS CSS Transitions with transition: all: property <hyphenate-character> from [initial] to ["e"] at (0.6) should be ["e"]
PASS CSS Transitions with transition: all: property <hyphenate-character> from [initial] to ["e"] at (1) should be ["e"]
PASS CSS Transitions with transition: all: property <hyphenate-character> from [initial] to ["e"] at (1.5) should be ["e"]
PASS CSS Animations: property <hyphenate-character> from [initial] to ["e"] at (-0.3) should be [initial]
PASS CSS Animations: property <hyphenate-character> from [initial] to ["e"] at (0) should be [initial]
PASS CSS Animations: property <hyphenate-character> from [initial] to ["e"] at (0.3) should be [initial]
FAIL CSS Animations: property <hyphenate-character> from [initial] to ["e"] at (0.5) should be ["e"] assert_equals: expected "\" e \" " but got "auto "
FAIL CSS Animations: property <hyphenate-character> from [initial] to ["e"] at (0.6) should be ["e"] assert_equals: expected "\" e \" " but got "auto "
FAIL CSS Animations: property <hyphenate-character> from [initial] to ["e"] at (1) should be ["e"] assert_equals: expected "\" e \" " but got "auto "
FAIL CSS Animations: property <hyphenate-character> from [initial] to ["e"] at (1.5) should be ["e"] assert_equals: expected "\" e \" " but got "auto "
PASS Web Animations: property <hyphenate-character> from [initial] to ["e"] at (-0.3) should be [initial]
PASS Web Animations: property <hyphenate-character> from [initial] to ["e"] at (0) should be [initial]
PASS Web Animations: property <hyphenate-character> from [initial] to ["e"] at (0.3) should be [initial]
FAIL Web Animations: property <hyphenate-character> from [initial] to ["e"] at (0.5) should be ["e"] assert_equals: expected "\" e \" " but got "auto "
FAIL Web Animations: property <hyphenate-character> from [initial] to ["e"] at (0.6) should be ["e"] assert_equals: expected "\" e \" " but got "auto "
FAIL Web Animations: property <hyphenate-character> from [initial] to ["e"] at (1) should be ["e"] assert_equals: expected "\" e \" " but got "auto "
FAIL Web Animations: property <hyphenate-character> from [initial] to ["e"] at (1.5) should be ["e"] assert_equals: expected "\" e \" " but got "auto "
FAIL CSS Transitions: property <hyphenate-limit-chars> from [initial] to [10] at (-0.3) should be [10] assert_true: 'from' value should be supported expected true got false
FAIL CSS Transitions: property <hyphenate-limit-chars> from [initial] to [10] at (0) should be [10] assert_true: 'from' value should be supported expected true got false
FAIL CSS Transitions: property <hyphenate-limit-chars> from [initial] to [10] at (0.3) should be [10] assert_true: 'from' value should be supported expected true got false
FAIL CSS Transitions: property <hyphenate-limit-chars> from [initial] to [10] at (0.5) should be [10] assert_true: 'from' value should be supported expected true got false
FAIL CSS Transitions: property <hyphenate-limit-chars> from [initial] to [10] at (0.6) should be [10] assert_true: 'from' value should be supported expected true got false
FAIL CSS Transitions: property <hyphenate-limit-chars> from [initial] to [10] at (1) should be [10] assert_true: 'from' value should be supported expected true got false
FAIL CSS Transitions: property <hyphenate-limit-chars> from [initial] to [10] at (1.5) should be [10] assert_true: 'from' value should be supported expected true got false
FAIL CSS Transitions with transition: all: property <hyphenate-limit-chars> from [initial] to [10] at (-0.3) should be [10] assert_true: 'from' value should be supported expected true got false
FAIL CSS Transitions with transition: all: property <hyphenate-limit-chars> from [initial] to [10] at (0) should be [10] assert_true: 'from' value should be supported expected true got false
FAIL CSS Transitions with transition: all: property <hyphenate-limit-chars> from [initial] to [10] at (0.3) should be [10] assert_true: 'from' value should be supported expected true got false
FAIL CSS Transitions with transition: all: property <hyphenate-limit-chars> from [initial] to [10] at (0.5) should be [10] assert_true: 'from' value should be supported expected true got false
FAIL CSS Transitions with transition: all: property <hyphenate-limit-chars> from [initial] to [10] at (0.6) should be [10] assert_true: 'from' value should be supported expected true got false
FAIL CSS Transitions with transition: all: property <hyphenate-limit-chars> from [initial] to [10] at (1) should be [10] assert_true: 'from' value should be supported expected true got false
FAIL CSS Transitions with transition: all: property <hyphenate-limit-chars> from [initial] to [10] at (1.5) should be [10] assert_true: 'from' value should be supported expected true got false
FAIL CSS Animations: property <hyphenate-limit-chars> from [initial] to [10] at (-0.3) should be [initial] assert_true: 'from' value should be supported expected true got false
FAIL CSS Animations: property <hyphenate-limit-chars> from [initial] to [10] at (0) should be [initial] assert_true: 'from' value should be supported expected true got false
FAIL CSS Animations: property <hyphenate-limit-chars> from [initial] to [10] at (0.3) should be [initial] assert_true: 'from' value should be supported expected true got false
FAIL CSS Animations: property <hyphenate-limit-chars> from [initial] to [10] at (0.5) should be [10] assert_true: 'from' value should be supported expected true got false
FAIL CSS Animations: property <hyphenate-limit-chars> from [initial] to [10] at (0.6) should be [10] assert_true: 'from' value should be supported expected true got false
FAIL CSS Animations: property <hyphenate-limit-chars> from [initial] to [10] at (1) should be [10] assert_true: 'from' value should be supported expected true got false
FAIL CSS Animations: property <hyphenate-limit-chars> from [initial] to [10] at (1.5) should be [10] assert_true: 'from' value should be supported expected true got false
FAIL Web Animations: property <hyphenate-limit-chars> from [initial] to [10] at (-0.3) should be [initial] assert_true: 'from' value should be supported expected true got false
FAIL Web Animations: property <hyphenate-limit-chars> from [initial] to [10] at (0) should be [initial] assert_true: 'from' value should be supported expected true got false
FAIL Web Animations: property <hyphenate-limit-chars> from [initial] to [10] at (0.3) should be [initial] assert_true: 'from' value should be supported expected true got false
FAIL Web Animations: property <hyphenate-limit-chars> from [initial] to [10] at (0.5) should be [10] assert_true: 'from' value should be supported expected true got false
FAIL Web Animations: property <hyphenate-limit-chars> from [initial] to [10] at (0.6) should be [10] assert_true: 'from' value should be supported expected true got false
FAIL Web Animations: property <hyphenate-limit-chars> from [initial] to [10] at (1) should be [10] assert_true: 'from' value should be supported expected true got false
FAIL Web Animations: property <hyphenate-limit-chars> from [initial] to [10] at (1.5) should be [10] assert_true: 'from' value should be supported expected true got false
PASS CSS Transitions: property <hyphens> from [initial] to [auto] at (-0.3) should be [auto]
PASS CSS Transitions: property <hyphens> from [initial] to [auto] at (0) should be [auto]
PASS CSS Transitions: property <hyphens> from [initial] to [auto] at (0.3) should be [auto]
PASS CSS Transitions: property <hyphens> from [initial] to [auto] at (0.5) should be [auto]
PASS CSS Transitions: property <hyphens> from [initial] to [auto] at (0.6) should be [auto]
PASS CSS Transitions: property <hyphens> from [initial] to [auto] at (1) should be [auto]
PASS CSS Transitions: property <hyphens> from [initial] to [auto] at (1.5) should be [auto]
PASS CSS Transitions with transition: all: property <hyphens> from [initial] to [auto] at (-0.3) should be [auto]
PASS CSS Transitions with transition: all: property <hyphens> from [initial] to [auto] at (0) should be [auto]
PASS CSS Transitions with transition: all: property <hyphens> from [initial] to [auto] at (0.3) should be [auto]
PASS CSS Transitions with transition: all: property <hyphens> from [initial] to [auto] at (0.5) should be [auto]
PASS CSS Transitions with transition: all: property <hyphens> from [initial] to [auto] at (0.6) should be [auto]
PASS CSS Transitions with transition: all: property <hyphens> from [initial] to [auto] at (1) should be [auto]
PASS CSS Transitions with transition: all: property <hyphens> from [initial] to [auto] at (1.5) should be [auto]
PASS CSS Animations: property <hyphens> from [initial] to [auto] at (-0.3) should be [initial]
PASS CSS Animations: property <hyphens> from [initial] to [auto] at (0) should be [initial]
PASS CSS Animations: property <hyphens> from [initial] to [auto] at (0.3) should be [initial]
PASS CSS Animations: property <hyphens> from [initial] to [auto] at (0.5) should be [auto]
PASS CSS Animations: property <hyphens> from [initial] to [auto] at (0.6) should be [auto]
PASS CSS Animations: property <hyphens> from [initial] to [auto] at (1) should be [auto]
PASS CSS Animations: property <hyphens> from [initial] to [auto] at (1.5) should be [auto]
PASS Web Animations: property <hyphens> from [initial] to [auto] at (-0.3) should be [initial]
PASS Web Animations: property <hyphens> from [initial] to [auto] at (0) should be [initial]
PASS Web Animations: property <hyphens> from [initial] to [auto] at (0.3) should be [initial]
PASS Web Animations: property <hyphens> from [initial] to [auto] at (0.5) should be [auto]
PASS Web Animations: property <hyphens> from [initial] to [auto] at (0.6) should be [auto]
PASS Web Animations: property <hyphens> from [initial] to [auto] at (1) should be [auto]
PASS Web Animations: property <hyphens> from [initial] to [auto] at (1.5) should be [auto]
